backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx>=0.26.0
aiohttp>=3.9.0
websockets>=12.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.5.3
pydantic-settings>=2.4.0

# LLM & Observability
litellm==1.35.0
openai>=1.40.0,<2.0.0
sse-starlette==1.8.2
aiomysql==0.2.0

# Data Cloud - Database Connectors (SQLAlchemy 기반)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0  # PostgreSQL 동기 드라이버 (스키마 리플렉션용)
pymysql>=1.1.0  # MariaDB/MySQL 동기 드라이버 (스키마 리플렉션용)
cryptography>=42.0.0

# ClickHouse SQLAlchemy dialect
clickhouse-connect>=0.7.0
clickhouse-sqlalchemy>=0.3.0  # SQLAlchemy dialect for ClickHouse

# Oracle SQLAlchemy dialect
oracledb>=2.0.0  # Oracle DB 드라이버 (cx_Oracle 후속)

# SAP HANA SQLAlchemy dialect
hdbcli>=2.0.0  # SAP HANA 드라이버
sqlalchemy-hana>=0.5.0  # SQLAlchemy dialect for SAP HANA

# Databricks SQL Connector
databricks-sql-connector>=3.0.0  # Databricks 드라이버

# LangChain & LangGraph (Text-to-SQL Agent)
langchain>=0.3.15
langchain-core>=1.2.0
langchain-community>=0.3.15
langchain-openai>=0.3.0  # OpenAI adapter for LangChain (Opus Single Agent)
langgraph>=0.2.60

# OpenTelemetry (Observability)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

# Text-to-SQL utilities
pandas>=2.0.0
PyYAML>=6.0.0
sqlparse>=0.4.0
tabulate>=0.9.0

# MCP stdio support
mcp>=0.9.0  # Model Context Protocol SDK
GitPython>=3.1.0  # Git operations

# Slide Studio
python-pptx>=0.6.21  # PowerPoint generation
json-repair>=0.7.0  # JSON repair for LLM output
Pillow>=10.0.0  # Image processing
playwright>=1.40.0  # Browser automation for advanced verification